chore(sdk): split out relay server so it can be shared with yea-wandb (#4837)
