+++ mypy.ini
@@ -11,6 +11,9 @@ exclude = (?x)(
 [mypy-wandb.*]
 ignore_errors = True
 
+[mypy-wandb.testing.*]
+ignore_errors = True
+
 [mypy-wandb.env]
 ignore_errors = False
 disallow_incomplete_defs = True
