+++ tox.ini
@@ -16,7 +16,7 @@ envlist=
 
 [base]
 setenv =
-    YEA_WANDB_VERSION = 0.9.1
+    YEA_WANDB_VERSION = 0.9.5
     ; Setting low network buffer so that we exercise flow control logic
     WANDB__NETWORK_BUFFER = 1000
